requirements.txt
streamlit>=1.30.0
openai>=1.10.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
python-docx>=1.0.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
numpy>=1.24.0
pytest>=7.0.0
